+++ main.py
@@ -1,3 +1,4 @@
+import asyncio
 import os
 import warnings
 
@@ -5,6 +6,7 @@ import discord
 from dotenv import load_dotenv
 
 from functions import *
+from utilities import *
 
 
 # TODO
@@ -40,8 +42,15 @@ async def on_ready():
     # else:
     #     warnings.warn('Bot not connected to verbose server.')
 
+    # Status
+    await client.change_presence(status=discord.Status.online, activity=discord.Game('🎵some tunes🎵'))
+
     # Setup
     client.queues = {}
+    loop = asyncio.new_event_loop()
+    asyncio.set_event_loop(loop)
+    # loop.run_forever()
+
     print('Bot is configured, ready to operate.')
 
 
@@ -53,50 +62,53 @@ async def on_message(message):
     if not command[0] == '!': return
     command = command[1:].lower()
 
+    # Get loop
+    loop = asyncio.get_running_loop()
+
     # Connect to voice channel
     if command == 'join':
         await message.add_reaction('👍')
-        await join(client, message)
+        loop.create_task(type_during(message, join(client, message)))
 
     # Disconnect
     elif command == 'leave':
         await message.add_reaction('👍')
-        await disconnect(client, message)
+        loop.create_task(type_during(message, disconnect(client, message)))
 
     # Play music
     elif command == 'play':
         await message.add_reaction('👍')
-        await play(client, message)
+        loop.create_task(type_during(message, play(client, message)))
 
     # Skip current song
     elif command == 'skip':
         await message.add_reaction('👍')
-        await skip(client, message)
+        loop.create_task(type_during(message, skip(client, message)))
 
     # Pause current song
     elif command == 'pause':
         await message.add_reaction('👍')
-        await pause(client, message)
+        loop.create_task(type_during(message, pause(client, message)))
 
     # Resume current song
     elif command == 'resume':
         await message.add_reaction('👍')
-        await resume(client, message)
+        loop.create_task(type_during(message, resume(client, message)))
 
     # Preview current song
     elif command == 'playing' or command == 'current':
         await message.add_reaction('👍')
-        await previewCurrent(client, message)
+        loop.create_task(type_during(message, previewCurrent(client, message)))
 
     # Preview next song
     elif command == 'next':
         await message.add_reaction('👍')
-        await previewNext(client, message)
+        loop.create_task(type_during(message, previewNext(client, message)))
 
     # Preview queue
     elif command == 'queue':
         await message.add_reaction('👍')
-        await previewQueue(client, message)
+        loop.create_task(type_during(message, previewQueue(client, message)))
 
     else:
         await message.add_reaction('❌')